Move NewItemMenu from common into nav
