+++ frontend/src/metabase/nav/components/NewItemMenu/analytics.ts
